requirements.txt
datasets
deepspeed
gputil
torch==1.11.0+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
pytorch-lightning==1.6.5
tensorboard
transformers
accelerate
openai
gpustat